Fix CodeQL integer multiplication allocation alerts in player.c

+++ player.c
@@ -3792,7 +3792,7 @@ void *player_thread_func(void *arg) {
               die("Failed to allocate memory for the transition buffer.");
             // size change
             conn->outbuf =
-                malloc(sps_format_sample_size(
+                malloc((size_t)sps_format_sample_size(
                            FORMAT_FROM_ENCODED_FORMAT(config.current_output_configuration)) *
                        CHANNELS_FROM_ENCODED_FORMAT(config.current_output_configuration) *
                        ((inframe->length) * conn->output_sample_ratio + INTERPOLATION_LIMIT));
@@ -4403,9 +4403,10 @@ void *player_thread_func(void *arg) {
                     skipping_frames_at_start_of_play = 0;
                     int64_t gap = -gap_to_fix;
                     void *silence = malloc(
-                        sps_format_sample_size(
+                        (size_t)sps_format_sample_size(
                             FORMAT_FROM_ENCODED_FORMAT(config.current_output_configuration)) *
-                        CHANNELS_FROM_ENCODED_FORMAT(config.current_output_configuration) * gap);
+                        CHANNELS_FROM_ENCODED_FORMAT(config.current_output_configuration) *
+                        (size_t)gap);
                     if (silence == NULL) {
                       debug(1, "Failed to allocate memory for a silent gap.");
                     } else {
